% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx.R
\name{hdx_dataset}
\alias{hdx_dataset}
\title{Get a specific HDX dataset by ID or name}
\usage{
hdx_dataset(id)
}
\arguments{
\item{id}{Dataset ID or name slug}
}
\value{
tibble: id, name, title, organization, metadata_created,
resource_name, resource_url, resource_format, resource_size (one row per resource)
}
\description{
Get a specific HDX dataset by ID or name
}
