% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx.R
\name{hdx_organizations}
\alias{hdx_organizations}
\title{List HDX organizations}
\usage{
hdx_organizations(limit = 20, offset = 0)
}
\arguments{
\item{limit}{Max results (default 20)}

\item{offset}{Offset for pagination (default 0)}
}
\value{
tibble: id, name, title, description, package_count
}
\description{
List HDX organizations
}
